chore/java_email_app: refactor typo
Renamed package modal to model
