+++ email_app_java/src/main/java/net/maidsafe/example/mail/controller/AppController.java
@@ -3,8 +3,8 @@ package net.maidsafe.example.mail.controller;
 import java.util.List;
 import net.maidsafe.example.mail.dao.IMessagingDao;
 import net.maidsafe.example.mail.dao.LauncherAPI;
-import net.maidsafe.example.mail.modal.Message;
-import net.maidsafe.example.mail.modal.Result;
+import net.maidsafe.example.mail.model.Message;
+import net.maidsafe.example.mail.model.Result;
 import net.maidsafe.example.mail.util.ServiceHandler;
 
 /**
