+++ email_app_java/src/main/java/net/maidsafe/example/mail/dao/IRestAPI.java
@@ -1,17 +1,17 @@
 package net.maidsafe.example.mail.dao;
 
-import net.maidsafe.example.mail.modal.AppConfig;
-import net.maidsafe.example.mail.modal.AppendableDataCreateRequest;
-import net.maidsafe.example.mail.modal.AppendableDataDataId;
-import net.maidsafe.example.mail.modal.AppendableDataMetadata;
-import net.maidsafe.example.mail.modal.AuthRequest;
-import net.maidsafe.example.mail.modal.AuthResponse;
-import net.maidsafe.example.mail.modal.HandleIdResponse;
-import net.maidsafe.example.mail.modal.ReaderInfo;
-import net.maidsafe.example.mail.modal.StructuredDataCreateRequest;
-import net.maidsafe.example.mail.modal.StructuredDataDataId;
-import net.maidsafe.example.mail.modal.StructuredDataMetadata;
-import net.maidsafe.example.mail.modal.StructuredDataUpdateRequest;
+import net.maidsafe.example.mail.model.AppConfig;
+import net.maidsafe.example.mail.model.AppendableDataCreateRequest;
+import net.maidsafe.example.mail.model.AppendableDataDataId;
+import net.maidsafe.example.mail.model.AppendableDataMetadata;
+import net.maidsafe.example.mail.model.AuthRequest;
+import net.maidsafe.example.mail.model.AuthResponse;
+import net.maidsafe.example.mail.model.HandleIdResponse;
+import net.maidsafe.example.mail.model.ReaderInfo;
+import net.maidsafe.example.mail.model.StructuredDataCreateRequest;
+import net.maidsafe.example.mail.model.StructuredDataDataId;
+import net.maidsafe.example.mail.model.StructuredDataMetadata;
+import net.maidsafe.example.mail.model.StructuredDataUpdateRequest;
 import okhttp3.RequestBody;
 import okhttp3.ResponseBody;
 import retrofit2.Call;
