+++ email_app_java/src/main/java/net/maidsafe/example/mail/model/AppendableDataDataId.java
@@ -1,4 +1,4 @@
-package net.maidsafe.example.mail.modal;
+package net.maidsafe.example.mail.model;
 
 /**
  *
